Add script for generating videos with ICface
Requires encodings to be computed first.
This script is a bit hacky.  The only alternatives I could see where
to refactor ICface or to manipulate the fields of various ICface classes.
The former would be very time consuming and the latter, would rely on
behavior that could change with future versions.